requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-daisyui==0.2.0
Django==4.2.8
django-celery-results==2.5.1
django-crispy-forms==2.1
django-ninja==1.1.0
kombu==5.3.4
numpy==1.26.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
uWSGI==2.0.23
